Clean up Node builder
Removed `build_pyfunc`, since it is easier just to make a SimPyFunc.
